requirements.txt
# QuoteSnap Backend Requirements
# Core Flask framework and extensions
flask==3.0.0
flask-cors==4.0.0

# Google APIs
google-api-python-client==2.110.0
google-auth-httplib2==0.1.1
google-auth-oauthlib==1.1.0

# OpenAI for AI extraction
openai==1.3.8

# Excel processing
openpyxl==3.1.2
pandas==2.1.4

# Environment and configuration
python-dotenv==1.0.0

# HTTP requests
requests==2.31.0

# Email validation
email-validator==2.1.0

# Additional utilities that may be needed
# Uncomment as required

# PDF processing (for Phase 2)
# PyPDF2==3.0.1
# reportlab==4.0.8

# WhatsApp API (for Phase 2)
# twilio==8.10.3

# Vector database (for Phase 2)
# pinecone-client==2.2.4
# qdrant-client==1.7.0

# HTML processing
# beautifulsoup4==4.12.2
# lxml==4.9.3

# Task scheduling
# celery==5.3.4
# redis==5.0.1

# Testing
# pytest==7.4.3
# pytest-flask==1.3.0